refactor(prefigure): extract shared formula helpers into formulaUtils
Move `astToExpressionString`, `rewriteExpressionVariable`, and
`parseFormulaDefinition` (with its `ParsedFormulaDefinition` interface) out
of `components/curve.ts` and `components/regionBetweenCurves.ts` into a new
`formulaUtils.ts` module. Both component converters now import the shared
implementations, removing the small block of duplicated helpers introduced
when the regionBetweenCurves converter was added.
